Django_jianlong/requirements.txt
absl-py==0.9.0
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
astor==0.8.1
astunparse==1.6.3
audioread==2.1.8
baidu-aip==2.2.18.0
billiard==3.3.0.23
bleach==1.5.0
cachetools==4.1.0
celery==3.1.26.post2
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
Django==3.0.7
django-celery==3.3.1
future==0.18.2
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
importlib-metadata==1.6.1
joblib==0.15.1
Keras==2.1.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kombu==3.0.37
librosa==0.6.3
llvmlite==0.30.0
Markdown==3.0.1
mock==3.0.5
numba==0.46.0
numpy==1.16.3
oauthlib==3.1.0
opt-einsum==3.2.1
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2018.9
PyYAML==5.1.2
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-learn==0.21.3
scipy==1.4.1
six==1.12.0
sklearn==0.0
SoundFile==0.10.2
sqlparse==0.3.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.5.0
tensorflow-estimator==2.2.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
urllib3==1.24.1
vine==1.3.0
Werkzeug==0.15.4
wrapt==1.12.1
zipp==3.1.0
